+++ src/api/engine/character/fetchCharacterByNick.ts
@@ -1,9 +1,6 @@
 import { Character } from "../../../interfaces/character.interface";
-import { logger } from "../../../config/internal/logger";
-import { Result, Error } from "../../../types/result.type";
-import axios from "axios";
 import { engineApiUrl } from "../config";
-import { undefinedError } from "../../../config/internal/error";
+import { requestHandler } from "../../requestHandler";
 
 /**
  * Fetch character by unique nick
@@ -17,29 +14,9 @@ import { undefinedError } from "../../../config/internal/error";
  * @param nick - Character nick
  * @returns Tuple of [Error, Character]
  */
-export const fetchCharacterByNick = async (
-  nick: string
-): Promise<Result<Character>> => {
-  logger.write("FETCH_CHARACTER_BY_NICK", { nick });
-
-  try {
-    const response = await axios(`${engineApiUrl}/character/nick/${nick}`, {
-      method: "GET"
-    });
-
-    const responseData = response.data;
-
-    logger.write("FETCHED_CHARACTER_BY_NICK", { nick, responseData });
-
-    return [null, responseData as Character];
-  } catch (error) {
-    const errorData = (error?.response?.data as Error) || undefinedError;
-
-    logger.write("FETCH_CHARACTER_BY_NICK_ERR", {
-      nick,
-      error: errorData?.statusCode
-    });
-
-    return [errorData, null];
-  }
-};
+export const fetchCharacterByNick = (nick: string) =>
+  requestHandler<Character>(`${engineApiUrl}/character/nick/${nick}`, "GET", {
+    start: "FETCH_CHARACTER_BY_NICK",
+    end: "FETCHED_CHARACTER_BY_NICK",
+    err: "FETCH_CHARACTER_BY_NICK_ERR"
+  });