+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - AG-26, Get on track a whole Angotia sdk
 - AG-31, Fix error handling when `Engine API` is turned off
 - AG-32, Extend user config character section
+- AG-33, Create request handler